requirements.txt
streamlit
loguru
python-dotenv
requests
openai


